+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add support for Inbound and Outbound streaming RPCs using gRPC and Protobuf.
 - Add support for creating peer choosers through config with PeerChooserSpec.
 
+### Fixed
+- Tchannel inbound response errors are now properly mapped from yarpc errors.
+
 
 ## [1.26.1] - 2017-12-21
 ### Removed